recommendations/requirements.txt
uvicorn==0.16.0
fastapi==0.70.0
pandas==1.3.4
python-dotenv==0.19.2
